LPS-56785 Reorganize packages for consistency
