+++ modules/apps/microblogs/microblogs-web/src/com/liferay/microblogs/web/notifications/MicroblogsReplyToTaggedUserNotificationDefinition.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.microblogs.web.microblogs.notifications;
+package com.liferay.microblogs.web.notifications;
 
 import com.liferay.microblogs.constants.MicroblogsPortletKeys;
 import com.liferay.microblogs.model.MicroblogsEntryConstants;
@@ -30,14 +30,15 @@ import org.osgi.service.component.annotations.Component;
 	property = {"javax.portlet.name=" + MicroblogsPortletKeys.MICROBLOGS},
 	service = UserNotificationDefinition.class
 )
-public class MicroblogsReplyUserNotificationDefinition
+public class MicroblogsReplyToTaggedUserNotificationDefinition
 	extends UserNotificationDefinition {
 
-	public MicroblogsReplyUserNotificationDefinition() {
+	public MicroblogsReplyToTaggedUserNotificationDefinition() {
 		super(
 			MicroblogsPortletKeys.MICROBLOGS, 0,
-			MicroblogsEntryConstants.NOTIFICATION_TYPE_REPLY,
-			"receive-a-notification-when-someone-comments-on-your-microblog");
+			MicroblogsEntryConstants.NOTIFICATION_TYPE_REPLY_TO_TAGGED,
+			"receive-a-notification-when-someone-comments-on-a-microblog-you-" +
+				"are-tagged-in");
 
 		addUserNotificationDeliveryType(
 			new UserNotificationDeliveryType(
