+++ modules/apps/microblogs/microblogs-web/src/com/liferay/microblogs/web/notifications/MicroblogsReplyUserNotificationDefinition.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.microblogs.web.microblogs.notifications;
+package com.liferay.microblogs.web.notifications;
 
 import com.liferay.microblogs.constants.MicroblogsPortletKeys;
 import com.liferay.microblogs.model.MicroblogsEntryConstants;
@@ -30,15 +30,14 @@ import org.osgi.service.component.annotations.Component;
 	property = {"javax.portlet.name=" + MicroblogsPortletKeys.MICROBLOGS},
 	service = UserNotificationDefinition.class
 )
-public class MicroblogsReplyToTaggedUserNotificationDefinition
+public class MicroblogsReplyUserNotificationDefinition
 	extends UserNotificationDefinition {
 
-	public MicroblogsReplyToTaggedUserNotificationDefinition() {
+	public MicroblogsReplyUserNotificationDefinition() {
 		super(
 			MicroblogsPortletKeys.MICROBLOGS, 0,
-			MicroblogsEntryConstants.NOTIFICATION_TYPE_REPLY_TO_TAGGED,
-			"receive-a-notification-when-someone-comments-on-a-microblog-you-" +
-				"are-tagged-in");
+			MicroblogsEntryConstants.NOTIFICATION_TYPE_REPLY,
+			"receive-a-notification-when-someone-comments-on-your-microblog");
 
 		addUserNotificationDeliveryType(
 			new UserNotificationDeliveryType(
