+++ modules/apps/microblogs/microblogs-web/src/com/liferay/microblogs/web/portlet/MicroblogsPortlet.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.microblogs.web.microblogs.portlet;
+package com.liferay.microblogs.web.portlet;
 
 import com.liferay.microblogs.constants.MicroblogsPortletKeys;
